% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_stats.R
\name{deviance.jglmm}
\alias{deviance.jglmm}
\title{Model Deviance}
\usage{
\method{deviance}{jglmm}(object, ...)
}
\arguments{
\item{object}{An object of class `jglmm`, as returned by `jglmm()`.}

\item{...}{Optional additional arguments, currently none are used.}
}
\value{
A numeric giving the deviance extracted from the fitted model.
}
\description{
Model Deviance
}
\examples{
\dontrun{
jglmm_setup()
lm1 <- jglmm(Reaction ~ Days + (Days | Subject), lme4::sleepstudy)
deviance(lm1)
}
}
