% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_stats.R
\name{extractAIC.jglmm}
\alias{extractAIC.jglmm}
\title{Extract AIC from a Fitted Model}
\usage{
\method{extractAIC}{jglmm}(x, k = 2)
}
\arguments{
\item{x}{An object of class `jglmm`, as returned by `jglmm`.}

\item{k}{Numeric specifying the 'weight' of the
\emph{degrees of freedom} part in the AIC formula.}
}
\value{
A numeric vector of length 2, with first and second elements giving

   `df` the '\strong{d}egrees of \strong{f}reedom' for the fitted model in
   `x`.

   `AIC` the (generalized) Akaike Information Criterion for the fitted model
   in `x`.
}
\description{
Extract AIC from a Fitted Model
}
\examples{
\dontrun{
jglmm_setup()
lm1 <- jglmm(Reaction ~ Days + (Days | Subject), lme4::sleepstudy)
aic <- extractAIC(lm1)
bic <- extractAIC(lm1, k = log(nobs(lm1)))
}
}
