% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{fitted.jglmm}
\alias{fitted.jglmm}
\title{Extract model fitted values}
\usage{
\method{fitted}{jglmm}(x)
}
\arguments{
\item{x}{An object of class `jglmm`, as returned by `jglmm()`.}
}
\value{
Vector of fitted values extracted from the model.
}
\description{
Extract the fitted values from a `jglmm` object.
}
\examples{
\dontrun{
jglmm_setup()
cbpp <- dplyr::mutate(lme4::cbpp, prop = incidence / size)
gm <- jglmm(prop ~ period + (1 | herd), data = cbpp, family = "binomial",
            weights = cbpp$size)
fitted(gm)
}
}
