% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{fixef.jglmm}
\alias{fixef.jglmm}
\title{Extract fixed-effects estimates}
\usage{
\method{fixef}{jglmm}(object, ...)
}
\arguments{
\item{object}{An object of class `jglmm`, as returned by `jglmm()`.}

\item{...}{Optional additional arguments, currently none are used.}
}
\value{
A named numeric vector of fixed-effects estimates.
}
\description{
Extract the fixed-effects estimates from a `jglmm` object.
}
\examples{
\dontrun{
jglmm_setup()
lm1 <- jglmm(Reaction ~ Days + (Days | Subject), lme4::sleepstudy)
fixef(lm1)
}
}
