% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{hatvalues.jglmm}
\alias{hatvalues.jglmm}
\title{Diagonal elements of the hat matrix}
\usage{
\method{hatvalues}{jglmm}(x)
}
\arguments{
\item{x}{An object of class `jglmm`, as returned by `jglmm()`.}
}
\value{
A numeric vector containing the diagonal elements of the hat matrix.
}
\description{
Returns the values on the diagonal of the hat matrix, which is the matrix
that transforms the response vector (minus any offset) into the fitted values
(minus any offset). Note that this method should only be used for linear
mixed models. It is not clear if the hat matrix concept even makes sense for
generalized linear mixed models.
}
\examples{
\dontrun{
jglmm_setup()
lm1 <- jglmm(Reaction ~ Days + (Days | Subject), lme4::sleepstudy)
hatvalues(lm1)
}
}
