% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jglmm-package.R
\docType{package}
\name{jglmm-package}
\alias{jglmm-package}
\alias{_PACKAGE}
\title{jglmm: Generalized Mixed-Effects Models in Julia}
\description{
R package for interfacing with Julia's MixedModels library to fit generalized linear mixed-effects models, similar to the lme4 package in R (\url{http://dmbates.github.io/MixedModels.jl/latest/}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mikabr/jglmm}
  \item Report bugs at \url{http://github.com/mikabr/jglmm/issues}
}

}
\author{
\strong{Maintainer}: Mika Braginsky \email{mika.br@gmail.com}

}
\keyword{internal}
