% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R, R/extract.R
\name{jglmm}
\alias{jglmm}
\alias{print.jglmm}
\alias{summary.jglmm}
\title{Fitting Generalized Linear Mixed-Effects Models in Julia}
\usage{
jglmm(
  formula,
  data,
  family = "normal",
  link = NULL,
  weights = NULL,
  contrasts = NULL,
  REML = FALSE
)

\method{print}{jglmm}(x, ...)

\method{summary}{jglmm}(object, ...)
}
\arguments{
\item{formula}{A two-sided linear formula object describing both the
fixed-effects and random-effects part of the model, with the response on
the left of a ~ operator and the terms, separated by + operators, on the
right. Random-effects terms are distinguished by vertical bars ("|")
separating expressions for design matrices from grouping factors.}

\item{data}{A data frame containing the variables named in formula.}

\item{family}{(optional) The distribution family for the response variable
(defaults to "normal").}

\item{link}{(optional) The model link function (defaults to "identity").}

\item{weights}{(optional) A vector of prior case weights.}

\item{contrasts}{(optional) A named list mapping column names of categorical
variables in data to coding schemes (defaults to dummy coding all
categorical variables).}

\item{REML}{(optional) A logical indicating whether REML should be used
instead of maximum likelihood (defaults to \code{FALSE}).}

\item{x}{An object of class `jglmm`, as returned by `jglmm()`.}

\item{...}{Optional additional arguments, currently none are used.}

\item{object}{An object of class `jglmm`, as returned by `jglmm()`.}
}
\value{
An object of class `jglmm`.
}
\description{
Fitting Generalized Linear Mixed-Effects Models in Julia
}
\examples{
\dontrun{
jglmm_setup()
# linear model
lm1 <- jglmm(Reaction ~ Days + (Days | Subject), lme4::sleepstudy)

# logistic model
cbpp <- dplyr::mutate(lme4::cbpp, prop = incidence / size)
gm1 <- jglmm(prop ~ period + (1 | herd), data = cbpp, family = "binomial",
             weights = cbpp$size)
gm2 <- jglmm(prop ~ period + (1 | herd), data = cbpp, family = "binomial",
             weights = cbpp$size, contrasts = list(period = "effects"))
}
}
