% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jglmm.r
\name{jglmm_setup}
\alias{jglmm_setup}
\title{Set up Julia and required libraries}
\usage{
jglmm_setup(run_glmm_model = FALSE, quietly = FALSE)
}
\arguments{
\item{run_glmm_model}{(logical): run simple GLMM so JIT compilation gets done (speed up timing of subsequent models)}

\item{quietly}{print progress messages?}
}
\description{
Set up Julia and required libraries
}
