% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_stats.R
\name{logLik.jglmm}
\alias{logLik.jglmm}
\title{Extract Log-Likelihood}
\usage{
\method{logLik}{jglmm}(x)
}
\arguments{
\item{x}{An object of class `jglmm`, as returned by `jglmm`.}
}
\value{
An object of class \code{logLik}, a number with attribute \code{"df"}
  (\strong{d}egrees of \strong{f}reedom), giving the number of (estimated)
  parameters in the model.
}
\description{
Extract Log-Likelihood
}
\examples{
\dontrun{
jglmm_setup()
lm1 <- jglmm(Reaction ~ Days + (Days | Subject), lme4::sleepstudy)
logLik(lm1)
}
}
