% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_stats.R
\name{nobs.jglmm}
\alias{nobs.jglmm}
\title{Extract the Number of Observations from a Fit}
\usage{
\method{nobs}{jglmm}(object, ...)
}
\arguments{
\item{object}{An object of class `jglmm`, as returned by `jglmm()`.}

\item{...}{Optional additional arguments, currently none are used.}
}
\value{
A numeric giving the number of observations.
}
\description{
Extract the Number of Observations from a Fit
}
\examples{
\dontrun{
jglmm_setup()
lm1 <- jglmm(Reaction ~ Days + (Days | Subject), lme4::sleepstudy)
nobs(lm1)
}
}
