% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.jglmm}
\alias{predict.jglmm}
\title{Predictions from a model}
\usage{
\method{predict}{jglmm}(object, newdata = NULL, type = "link", allow.new.levels = FALSE, ...)
}
\arguments{
\item{object}{An object of class `jglmm`, as returned by `jglmm`.}

\item{newdata}{(optional) A dataframe of new data.}

\item{type}{(optional) A character string - either \code{"link"} (default),
where predictions are returned on the scale of the linear predictors, or
\code{"response"}, where predictions are returned on the scale of the
response variable.}

\item{allow.new.levels}{(optional) A logical indicating whether new random
effects levels in \code{newdata} are allowed. If \code{FALSE} (default),
such new values in \code{newdata} will trigger an error; if \code{TRUE},
then the prediction will use population-level values for data with
previously unobserved values.}

\item{...}{Optional additional arguments, currently none are used.}
}
\value{
A numeric vector of predicted values
}
\description{
Predictions from a model
}
\examples{
\dontrun{
jglmm_setup()
cbpp <- dplyr::mutate(lme4::cbpp, prop = incidence / size)
gm1 <- jglmm(prop ~ period + (1 | herd), data = cbpp, family = "binomial",
             weights = cbpp$size)
predict(gm1)
predict(gm1, type = "response")
newdata <- with(cbpp, expand.grid(period=unique(period), herd=unique(herd)))
predict(gm1, newdata)
}
}
