% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{ranef.jglmm}
\alias{ranef.jglmm}
\title{Extract the modes of the random effects}
\usage{
\method{ranef}{jglmm}(object, ...)
}
\arguments{
\item{object}{An object of class `jglmm`, as returned by `jglmm()`.}

\item{...}{Optional additional arguments, currently none are used.}
}
\value{
A list of data frames, one for each grouping factor for the random
  effects. The number of rows in the data frame is the number of levels of
  the grouping factor. The number of columns is the dimension of the random
  effect associated with each level of the factor. Each of the data frames
  has an attribute called "postVar", which contains an array for each
  random-effects term with the variance-covariance matrices for each level of
  the grouping factor.
}
\description{
Extract the conditional modes of the random effects from a `jglmm` object.
}
\examples{
\dontrun{
jglmm_setup()
cbpp <- dplyr::mutate(lme4::cbpp, prop = incidence / size)
gm <- jglmm(prop ~ period + (1 | herd), data = cbpp, family = "binomial",
            weights = cbpp$size)
ranef(gm)
}
}
