% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_ests.R
\name{ranef_jglmm}
\alias{ranef_jglmm}
\title{Extract the modes of the random effects}
\usage{
ranef_jglmm(x)
}
\arguments{
\item{x}{An object of class `jglmm`, as returned by `jglmm()`.}
}
\value{
A list of tibbles, one for each random effect group.
}
\description{
Extract the conditional modes of the random effects from a fitted `jglmm`
object.
}
\examples{
\dontrun{
jglmm_setup()
cbpp <- dplyr::mutate(lme4::cbpp, prop = incidence / size)
gm <- jglmm(prop ~ period + (1 | herd), data = cbpp, family = "binomial",
            weights = cbpp$size)
ranef_jglmm(gm)
}
}
