% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_ests.R
\name{sigma.jglmm}
\alias{sigma.jglmm}
\title{Extract residual standard deviation 'sigma'}
\usage{
\method{sigma}{jglmm}(x)
}
\arguments{
\item{x}{An object of class `jglmm`, as returned by `jglmm()`.}
}
\description{
Extract the estimated standard deviation of the errors, the "residual
standard deviation" (also misnamed the "residual standard error"), from a
fitted model.
}
\examples{
\dontrun{
jglmm_setup()
lm1 <- jglmm(Reaction ~ Days + (Days | Subject), lme4::sleepstudy)
sigma(lm1)
}
}
