% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{simulate.jglmm}
\alias{simulate.jglmm}
\title{Simulate responses from a model}
\usage{
\method{simulate}{jglmm}(x, newdata = NULL, nsim = 1)
}
\arguments{
\item{x}{An object of class `jglmm`, as returned by `jglmm`.}

\item{newdata}{(optional) A dataframe of new data.}

\item{nsim}{(optional) A positive integer indicating the number of responses
to simulate}
}
\value{
A numeric vector of simulated values
}
\description{
Simulate responses from a model
}
\examples{
\dontrun{
jglmm_setup()
cbpp <- dplyr::mutate(lme4::cbpp, prop = incidence / size)
gm1 <- jglmm(prop ~ period + (1 | herd), data = cbpp, family = "binomial",
             weights = cbpp$size)
simulate(gm1)
simulate(gm1, nsim = 5)
newdata <- with(cbpp, expand.grid(period=unique(period), herd=unique(herd)))
simulate(gm1, newdata)
}
}
