% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_ests.R
\name{vcov.jglmm}
\alias{vcov.jglmm}
\title{Calculate variance-covariance matrix for a fitted model object}
\usage{
\method{vcov}{jglmm}(x)
}
\arguments{
\item{x}{An object of class `jglmm`, as returned by `jglmm()`.}
}
\value{
A matrix of the estimated covariances between the parameter estimates
  in the linear or non-linear predictor of the model.
}
\description{
Calculate variance-covariance matrix for a fitted model object
}
\examples{
\dontrun{
jglmm_setup()
lm1 <- jglmm(Reaction ~ Days + (Days | Subject), lme4::sleepstudy)
vcov(lm1)
}
}
