% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jglmm.r
\name{jglmm_tidiers}
\alias{jglmm_tidiers}
\alias{tidy.jglmm}
\alias{augment.jglmm}
\title{Tidying methods for jglmm models}
\usage{
\method{tidy}{jglmm}(x)

\method{augment}{jglmm}(x)
}
\arguments{
\item{x}{An object of class `jglmm`, as returned by `jglmm`.}
}
\value{
`tidy` returns a tibble of fixed effect estimates

`augment` returns a tibble of the original data used to fit the model
  with an additional `.fitted` column containing the fitted response valuese.
}
\description{
Tidying methods for jglmm models
}
\examples{
\dontrun{
cbpp <- dplyr::mutate(lme4::cbpp, prop = incidence / size)
gm <- jglmm(prop ~ period + (1 | herd), data = cbpp, family = "binomial",
            weights = cbpp$size)
tidy(gm)
augment(gm)
}
}
