% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{simulate.jglmm}
\alias{simulate.jglmm}
\title{Simulate responses from a model}
\usage{
\method{simulate}{jglmm}(object, nsim = 1, seed = NULL, newdata = NULL, ...)
}
\arguments{
\item{object}{An object of class `jglmm`, as returned by `jglmm`.}

\item{nsim}{(optional) A positive integer indicating the number of responses
to simulate}

\item{seed}{Not currently used (see `simulate` generic).}

\item{newdata}{(optional) A dataframe of new data.}

\item{...}{Optional additional arguments, currently none are used.}
}
\value{
A numeric vector of simulated values
}
\description{
Simulate responses from a model
}
\examples{
\dontrun{
jglmm_setup()
cbpp <- dplyr::mutate(lme4::cbpp, prop = incidence / size)
gm1 <- jglmm(prop ~ period + (1 | herd), data = cbpp, family = "binomial",
             weights = cbpp$size)
simulate(gm1)
simulate(gm1, nsim = 5)
newdata <- with(cbpp, expand.grid(period=unique(period), herd=unique(herd)))
simulate(gm1, newdata = newdata)
}
}
