% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{vcov.jglmm}
\alias{vcov.jglmm}
\title{Calculate variance-covariance matrix for a fitted model object}
\usage{
\method{vcov}{jglmm}(object, ...)
}
\arguments{
\item{object}{An object of class `jglmm`, as returned by `jglmm()`.}

\item{...}{Optional additional arguments, currently none are used.}
}
\value{
A matrix of the estimated covariances between the parameter estimates
  in the linear or non-linear predictor of the model.
}
\description{
Extract the variance-covariance matrix of the main parameters from a
`jglmm` object.
}
\examples{
\dontrun{
jglmm_setup()
lm1 <- jglmm(Reaction ~ Days + (Days | Subject), lme4::sleepstudy)
vcov(lm1)
}
}
